+++ starkinfra/event/__event.py
@@ -16,6 +16,7 @@ from ..pixpullrequest.log.__log import _resource as _pixpullrequest_log_resource
 from ..issuingcard.log.__log import _resource as _issuingcard_log_resource
 from ..issuinginvoice.log.__log import _resource as _issuinginvoice_log_resource
 from ..issuingpurchase.log.__log import _resource as _issuingpurchase_log_resource
+from ..businessidentity.log.__log import _resource as _businessidentity_log_resource
 
 
 _resource_by_subscription = {
@@ -34,6 +35,7 @@ _resource_by_subscription = {
     "issuing-invoice": _issuinginvoice_log_resource,
     "issuing-purchase": _issuingpurchase_log_resource,
     "credit-note": _creditnote_log_resource,
+    "business-identity": _businessidentity_log_resource,
 }
 
 
@@ -47,7 +49,7 @@ class Event(Resource):
     - log [Log]: a Log object from one of the subscribed services (PixRequestLog, PixReversalLog)
     - created [datetime.datetime]: creation datetime for the notification Event. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     - is_delivered [bool]: true if the Event has been successfully delivered to the user url. ex: False
-    - subscription [string]: service that triggered this Event. Options: "pix-request.in", "pix-request.out", "pix-reversal.in", "pix-reversal.out", "pix-key", "pix-claim", "pix-infraction", "pix-chargeback", "pix-dispute", "pix-pull-subscription", "pix-pull-request", "issuing-card", "issuing-invoice", "issuing-purchase", "credit-note"
+    - subscription [string]: service that triggered this Event. Options: "pix-request.in", "pix-request.out", "pix-reversal.in", "pix-reversal.out", "pix-key", "pix-claim", "pix-infraction", "pix-chargeback", "pix-dispute", "pix-pull-subscription", "pix-pull-request", "issuing-card", "issuing-invoice", "issuing-purchase", "credit-note", "business-identity"
     - workspace_id [string]: ID of the Workspace that generated this Event. Mostly used when multiple Workspaces have Webhooks registered to the same endpoint. ex: "4545454545454545"
     """